better scripts + restructuring

+++ runs/rund.sh
@@ -1,35 +1,39 @@
-rm -rf $HOME/.config/hypr/
-cp -r $DEV_ENV/hyprdesk/ $HOME/.config/hypr/
+echo 'DEV_ENV: ' $DEV_ENV
 
-rm -rf $HOME/.config/waybar/
-cp -r $DEV_ENV/waybar/ $HOME/.config/waybar/
+if [ -z "$DEV_ENV" ]; then
+    echo "DEV_ENV is not set"
+    exit 1
+fi
 
-rm -rf $HOME/.config/nvim/
-cp -r $DEV_ENV/nvim $HOME/.config/nvim/
-git clone --depth 1 https://github.com/wbthomason/packer.nvim\
- ~/.local/share/nvim/site/pack/packer/start/packer.nvim
-mkdir ~/.local/scripts
-cp -r $DEV_ENV/runs/tmux-sessionizer ~/.local/scripts
+configs=("waybar", "nvim", "ghostty", "nvim". "qutebrowser", "rofi", "spicetify")
 
-rm -rf $HOME/.config/ghostty
-cp -r $DEV_ENV/ghostty $HOME/.config/ghostty
+for cfg in [configs]; do
+    rm -rf $HOME/.config/cfg/
+    cp -r $DEV_ENV/cfg $HOME/.config/cfg
+
+
+if [$1 = '--desktop']; then
+    rm -rf $HOME/.config/hypr/
+    cp -r $DEV_ENV/hyprdesk/ $HOME/.config/hypr/
+    cp $DEV_ENV/runs/tmux-sessionizer-desktop $HOME/.local/scripts/tmux-sessionizer
+else
+    rm -rf $HOME/.config/hypr/
+    cp -r $DEV_ENV/hyprtop/ $HOME/.config/hypr/
+    cp $DEV_ENV/runs/tmux-sessionizer $HOME/.local/scripts
+
+
+
+if [! -d $HOME/.local/scripts]
+    mkdir $HOME/.local/scripts
+
+
+git clone --depth 1 https://github.com/wbthomason/packer.nvim $HOME/.local/share/nvim/site/pack/packer/start/packer.nvim
+git clone https://github.com/tmux-plugins/tpm $HOME/.tmux/plugins/tpm
+
 
 rm $HOME/.zshrc $HOME/.tmux.conf
 cp $DEV_ENV/rcs/.zshrc $HOME
 cp $DEV_ENV/rcs/.tmux.conf $HOME
 
-mkdir ~/.local/scripts
-cp $DEV_ENV/runs/tmux-sessionizer-desktop ~/.local/scripts/tmux-sessionizer
-
-git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
-
-rm -rf $HOME/.config/qutebrowser/
-cp -r $DEV_ENV/qutebrowser $HOME/.config/qutebrowser/
-
-rm -rf $HOME/.config/rofi/
-cp -r $DEV_ENV/rofi $HOME/.config/rofi/
-
-rm -rf $HOME/.config/spicetify/
-cp -r $DEV_ENV/spicetify $HOME/.config/spicetify/
 hyprctl reload
 hyprshade toggle blue-light-filter
